Refactor LocalSave internals into focused modules
